requirements.txt
Flask==2.3.2
python-decouple==3.8
flask-SQLAlchemy==3.0.5
psycopg2==2.9.6
Flask-Migrate==4.0.4
flask-debugtoolbar==0.13.1
Flask-JWT-Extended==4.5.2